requirements.txt
PyQt6
pyqt6-tools
markdown
pyclamd
cryptography
pdfplumber
torch
torchvision
emoji
pyinstaller
llama-cpp-python
pytesseract
opencv-python
numpy
librosa
pandas
matplotlib
scikit-learn
requests
fastapi
uvicorn
gitpython
